xen/Rules.mk: fix build with CFLAGS from environment
When CFLAGS are passed from the environment, the first-level make
invocation will append -D__OBJECT_FILE__ to it, then call a second
make invocation, that will have those new CFLAGS in its environment,
but will also append -D__OBJECT_FILE__ to those.

Then, the compiler fails because __OBEJECT_FILE__ is defined twice.

Just undefine it before defining it again, as a *workaround* to this
issue.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>